End point reservas de citas 60%,end point Disponibilidad para verificar horarios dfisponibles de acuerdo al dia, reservas realizadas, horarios de trabajo y del profesional entre otros al 65%

+++ src/main/java/com/barberia/models/Servicio.java
@@ -36,7 +36,7 @@ public class Servicio {
     private Integer duracionMinutosAprox;
 
     @Column(nullable = false)
-    private Double precio;
+    private BigDecimal precio;
 
     @Column(nullable = false, name = "reg_estado")
     private Integer regEstado;
